Added sorted queue for sorting frames in the nex t changes (not done yet)
