+++ kernel/packages/voice-chat-codec/constants.ts
@@ -1,4 +1,4 @@
 export const OPUS_BITS_PER_SECOND = 24000
-export const OPUS_FRAME_SIZE_MS = 20
+export const OPUS_FRAME_SIZE_MS = 60
 
 export const OUTPUT_NODE_BUFFER_SIZE = 8192
